requirements/requirements_infer.txt
tqdm
six==1.16.0
easydict==1.9
ftfy
einops==0.6.1
decord==0.6.0
timm==0.4.12
torchlibrosa==0.0.9
librosa==0.9.2
omegaconf
matplotlib
diffusers==0.27.2
transformers==4.26.0
pillow
opencv-python 
numpy
pyyaml
accelerate
av
kornia
einops
matplotlib
pandas
pytorch-lightning==2.2.0
huggingface-hub==0.25.2
openpyxl